Move UpgradeInfo class to header and source files
