+++ lib/mapObjects/MiscObjects.cpp
@@ -32,6 +32,7 @@
 #include "../networkPacks/PacksForClient.h"
 #include "../networkPacks/PacksForClientBattle.h"
 #include "../networkPacks/StackLocation.h"
+#include "../lib/gameState/UpgradeInfo.h"
 
 #include <vstd/RNG.h>
 
@@ -1323,12 +1324,9 @@ void HillFort::fillUpgradeInfo(UpgradeInfo & info, const CStackInstance &stack) 
 
 	int costModifier = upgradeCostPercentage[index];
 
-	if(costModifier < 0) // upgrade not allowed
-		return;
-
 	for(const auto & nid : stack.getCreature()->upgrades)
 	{
-		info.addUpgrade(nid, stack.getType(), costModifier / 100);
+		info.addUpgrade(nid, stack.getType(), costModifier);
 	}
 }
